intermediate work
